+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Attempt fix for `MongoSocketRead` and `MongoSocketWrite` exceptions.
 - Detekt bug fixed - Detekt now works again!
 - Following on from that...fix _all_ Detekt issues that built up while it wasn't running.
+- Issue when  no public mod log channel set and using ban commands.
 
 ### Features